requirements.txt
Django==1.6.2
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.9.6
amqp==1.4.6
aniso8601==0.85
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.19
blinker==1.3
celery==3.1.17
colorama==0.3.2
django-rest-swagger==0.1.14
djangorestframework==2.4.2
itsdangerous==0.24
kombu==3.0.24
passlib==1.6.2
py==1.4.26
pytz==2014.10
six==1.8.0
wsgiref==0.1.2
django-debug-toolbar==1.3.2
sqlparse==0.1.18